Blurry bar quickfix

+++ libqtile/confreader.py
@@ -74,7 +74,7 @@ class Config:
     # Really we'd want to check this Any is libqtile.backend.wayland.ImportConfig, but
     # doing so forces the import, creating a hard dependency for wlroots.
     wl_input_rules: dict[str, Any] | None
-    wl_scale_factor: int
+    wl_scale_factor: float
 
     def __init__(self, file_path=None, **settings):
         """Create a Config() object from settings
